Refactor : commentResponse 에서 username 대신 user.profile.nick 을 넘겨주게 수정하였습니다.
