+++ src/main/java/com/sparta/newsfeed/common/GlobalExceptionController.java
@@ -1,14 +1,13 @@
 package com.sparta.newsfeed.common;
 
-import com.sparta.newsfeed.comment.exception.NoPrivilegesException;
-import com.sparta.newsfeed.comment.exception.NotFoundCommentException;
-import com.sparta.newsfeed.comment.exception.NotFoundPostException;
+import com.sparta.newsfeed.common.exception.comment.NoPrivilegesException;
+import com.sparta.newsfeed.common.exception.comment.NotFoundCommentException;
+import com.sparta.newsfeed.common.exception.comment.NotFoundPostException;
 import com.sparta.newsfeed.common.exception.InvalidUserException;
 import com.sparta.newsfeed.common.exception.NotFoundPostException;
 import com.sparta.newsfeed.common.exception.NotFoundUserException;
 import com.sparta.newsfeed.post.OnlyAuthorAccessException;
 import com.sparta.newsfeed.like.SelfLikeException;
-import com.sparta.newsfeed.user.*;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
