+++ src/main/java/com/sparta/newsfeed/common/exception/comment/NotFoundPostException.java
@@ -1,4 +1,4 @@
-package com.sparta.newsfeed.comment.exception;
+package com.sparta.newsfeed.common.exception.comment;
 
 public class NotFoundPostException  extends RuntimeException {
 	private static final String MESSAGE = "게시글이 존재하지 않습니다.";
